Add package skeleton from rstantools, stan model, basic R.
We will use the rstanarm/rstantools method for pre-building the model.
This means including code form rstanarm, and choosing the GPL-3 license.

The stan model is stored in exec, and should be compiled and ready to
go after installing the package.

R/model_fitting.R contains some basic functions for sampling from the stan
model and simulating fake data.